+++ package.json
@@ -43,6 +43,8 @@
   },
   "dependencies": {
     "invariant": "^2.2.0",
-    "lodash": "^3.x.x"
+    "lodash": "^3.x.x",
+    "react-devtools": "^2.0.12",
+    "ws": "^1.1.0"
   }
 }
